refactor: Move SettingsPageModule to GeneralSettingsModule in modules dir
