+++ Source/Plugins/Core/com.equella.core/js/tsrc/modules/GeneralSettingsModule.ts
